@@ -1,6 +1,6 @@
 {
   "name": "smartstock-sales",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "scripts": {
     "ng": "ng",
     "box": "yarn install && yarn start",
@@ -12,32 +12,32 @@
     "e2e": "ng e2e"
   },
   "dependencies": {
-    "@angular/animations": "~13.0.1",
-    "@angular/cdk": "~13.0.1",
-    "@angular/common": "~13.0.1",
-    "@angular/compiler": "~13.0.1",
-    "@angular/core": "~13.0.1",
-    "@angular/forms": "~13.0.1",
-    "@angular/material": "^13.0.1",
-    "@angular/material-moment-adapter": "^13.0.1",
-    "@angular/platform-browser": "~13.0.1",
-    "@angular/platform-browser-dynamic": "~13.0.1",
-    "@angular/router": "~13.0.1",
+    "@angular/animations": "^14.0.3",
+    "@angular/cdk": "^14.0.3",
+    "@angular/common": "^14.0.3",
+    "@angular/compiler": "^14.0.3",
+    "@angular/core": "^14.0.3",
+    "@angular/forms": "^14.0.3",
+    "@angular/material": "^14.0.3",
+    "@angular/material-moment-adapter": "^14.0.3",
+    "@angular/platform-browser": "^14.0.3",
+    "@angular/platform-browser-dynamic": "^14.0.3",
+    "@angular/router": "^14.0.3",
     "bfast": "6.0.1",
-    "comlink": "~4.3.1",
-    "crypto-hash": "~1.3.0",
-    "moment": "~2.29.2",
-    "ngx-infinite-scroll": "^10.0.1",
+    "comlink": "4.3.1",
+    "crypto-hash": "2.0.1",
+    "moment": "2.29.3",
+    "ngx-infinite-scroll": "13.0.2",
     "rxjs": "~7.4.0",
-    "smartstock-core": "1.0.1",
+    "smartstock-core": "1.0.2",
     "tslib": "^2.3.0",
-    "uuid": "~8.3.2",
+    "uuid": "8.3.2",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.2",
-    "@angular/cli": "~13.0.2",
-    "@angular/compiler-cli": "~13.0.1",
+    "@angular-devkit/build-angular": "^14.0.3",
+    "@angular/cli": "^14.0.3",
+    "@angular/compiler-cli": "^14.0.3",
     "@types/jasmine": "~3.6.2",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
@@ -49,10 +49,10 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^13.0.3",
+    "ng-packagr": "^14.0.2",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.4.4"
+    "typescript": "~4.7.4"
   }
 }
