@@ -8,7 +8,7 @@ import {
 } from "@angular/core";
 import { DeviceState } from "smartstock-core";
 import { RefundState } from "../states/refund.state";
-import { FormControl } from "@angular/forms";
+import { UntypedFormControl } from "@angular/forms";
 import { MatPaginator } from "@angular/material/paginator";
 
 @Component({
@@ -50,7 +50,7 @@ import { MatPaginator } from "@angular/material/paginator";
   styleUrls: ["../styles/refund-body.style.scss"]
 })
 export class RefundBodyHeaderMobileComponent implements OnInit {
-  dateControl = new FormControl(new Date());
+  dateControl = new UntypedFormControl(new Date());
 
   constructor(
     public readonly deviceState: DeviceState,
