@@ -117,119 +117,118 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  imports: [
-    CommonModule,
-    {
-      ngModule: RouterModule,
-      providers: [
+    imports: [
+        CommonModule,
         {
-          provide: ROUTES,
-          multi: true,
-          useValue: routes
-        }
-      ]
-    },
-    MatSortModule,
-    MatMenuModule,
-    MatSelectModule,
-    MatNativeDateModule,
-    MatSidenavModule,
-    MatFormFieldModule,
-    MatAutocompleteModule,
-    MatInputModule,
-    ReactiveFormsModule,
-    MatCardModule,
-    MatSlideToggleModule,
-    MatDividerModule,
-    MatIconModule,
-    MatTabsModule,
-    MatTableModule,
-    MatButtonModule,
-    MatPaginatorModule,
-    MatProgressSpinnerModule,
-    MatDatepickerModule,
-    MatSortModule,
-    MatTooltipModule,
-    MatToolbarModule,
-    MatBadgeModule,
-    MatRippleModule,
-    MatListModule,
-    ScrollingModule,
-    FormsModule,
-    CdkTableModule,
-    MatProgressBarModule,
-    MatChipsModule,
-    MatBottomSheetModule,
-    MatDialogModule,
-    MatExpansionModule,
-    MatSnackBarModule,
-    LibModule,
-    MatCheckboxModule,
-    InfiniteScrollModule
-  ],
-  declarations: [
-    CreateInvoicePage,
-    InvoiceProductTilesComponent,
-    InvoiceProductCardComponent,
-    AddToInvoiceCartDialogComponent,
-    AddToInvoiceCartFormComponent,
-    AddToInvoiceCartSheetComponent,
-    InvoiceProductListItemComponent,
-    InvoiceProductListComponent,
-    SaveInvoiceFormComponent,
-    SaveInvoiceSheetComponent,
-    SaveInvoiceDialogComponent,
-    InvoicesMobileComponent,
-    InvoiceCartComponent,
-    InvoicesDesktopComponent,
-    InvoicesTableOptionsComponent,
-    InvoicesTableComponent,
-    InvoiceDetailsModelComponent,
-    AddInvoicePaymentDialogComponent,
-    AddInvoicePaymentFormComponent,
-    InvoicePage,
-    CustomerActiveComponent,
-    CustomersTableComponent,
-    CustomersTableOptionsComponent,
-    AddToCartSheetComponent,
-    PeriodDateRangeComponent,
-    CreateRefundDialogComponent,
-    ReturnsDetailsComponent,
-    RefundsPage,
-    CustomersListComponent,
-    CustomersPage,
-    CreateCustomerComponent,
-    InvoiceDetailsComponent,
-    InfoDialogComponent,
-    CreateCreditorComponent,
-    RefundBodyHeaderMobileComponent,
-    RefundBodyListMobileComponent,
-    OrdersItemsComponent,
-    OrderPaymentStatusComponent,
-    IndexPage,
-    OrderPage,
-    WholePageComponent,
-    CartComponent,
-    ProductComponent,
-    RetailPageComponent,
-    SaleComponent,
-    CartPreviewComponent,
-    DeleteConfirmDialogComponent,
-    DialogCreateCustomerComponent,
-    SheetCreateCustomerComponent,
-    FedhaPipe,
-    MudaPipe,
-    DialogCashSaleCartOptionsComponent,
-    CashSaleCartOptionsComponent,
-    OrderListComponent,
-    DialogNewOrderComponent,
-    RefundBodyComponent,
-    RefundBodyMobileComponent,
-    RefundBodyHeaderComponent,
-    RefundBodyTableComponent,
-    AgoPipe
-  ],
-  entryComponents: []
+            ngModule: RouterModule,
+            providers: [
+                {
+                    provide: ROUTES,
+                    multi: true,
+                    useValue: routes
+                }
+            ]
+        },
+        MatSortModule,
+        MatMenuModule,
+        MatSelectModule,
+        MatNativeDateModule,
+        MatSidenavModule,
+        MatFormFieldModule,
+        MatAutocompleteModule,
+        MatInputModule,
+        ReactiveFormsModule,
+        MatCardModule,
+        MatSlideToggleModule,
+        MatDividerModule,
+        MatIconModule,
+        MatTabsModule,
+        MatTableModule,
+        MatButtonModule,
+        MatPaginatorModule,
+        MatProgressSpinnerModule,
+        MatDatepickerModule,
+        MatSortModule,
+        MatTooltipModule,
+        MatToolbarModule,
+        MatBadgeModule,
+        MatRippleModule,
+        MatListModule,
+        ScrollingModule,
+        FormsModule,
+        CdkTableModule,
+        MatProgressBarModule,
+        MatChipsModule,
+        MatBottomSheetModule,
+        MatDialogModule,
+        MatExpansionModule,
+        MatSnackBarModule,
+        LibModule,
+        MatCheckboxModule,
+        InfiniteScrollModule
+    ],
+    declarations: [
+        CreateInvoicePage,
+        InvoiceProductTilesComponent,
+        InvoiceProductCardComponent,
+        AddToInvoiceCartDialogComponent,
+        AddToInvoiceCartFormComponent,
+        AddToInvoiceCartSheetComponent,
+        InvoiceProductListItemComponent,
+        InvoiceProductListComponent,
+        SaveInvoiceFormComponent,
+        SaveInvoiceSheetComponent,
+        SaveInvoiceDialogComponent,
+        InvoicesMobileComponent,
+        InvoiceCartComponent,
+        InvoicesDesktopComponent,
+        InvoicesTableOptionsComponent,
+        InvoicesTableComponent,
+        InvoiceDetailsModelComponent,
+        AddInvoicePaymentDialogComponent,
+        AddInvoicePaymentFormComponent,
+        InvoicePage,
+        CustomerActiveComponent,
+        CustomersTableComponent,
+        CustomersTableOptionsComponent,
+        AddToCartSheetComponent,
+        PeriodDateRangeComponent,
+        CreateRefundDialogComponent,
+        ReturnsDetailsComponent,
+        RefundsPage,
+        CustomersListComponent,
+        CustomersPage,
+        CreateCustomerComponent,
+        InvoiceDetailsComponent,
+        InfoDialogComponent,
+        CreateCreditorComponent,
+        RefundBodyHeaderMobileComponent,
+        RefundBodyListMobileComponent,
+        OrdersItemsComponent,
+        OrderPaymentStatusComponent,
+        IndexPage,
+        OrderPage,
+        WholePageComponent,
+        CartComponent,
+        ProductComponent,
+        RetailPageComponent,
+        SaleComponent,
+        CartPreviewComponent,
+        DeleteConfirmDialogComponent,
+        DialogCreateCustomerComponent,
+        SheetCreateCustomerComponent,
+        FedhaPipe,
+        MudaPipe,
+        DialogCashSaleCartOptionsComponent,
+        CashSaleCartOptionsComponent,
+        OrderListComponent,
+        DialogNewOrderComponent,
+        RefundBodyComponent,
+        RefundBodyMobileComponent,
+        RefundBodyHeaderComponent,
+        RefundBodyTableComponent,
+        AgoPipe
+    ]
 })
 export class SalesModule {
   constructor(private readonly salesNav: SalesNavigationService) {
